Split `gradbench` crate into multiple modules
